project_1/advanced_rag/requirements.txt
sentence-transformers
faiss-cpu
groq
numpy
pandas
matplotlib
rank-bm25
